Restore required nullable port coercions

+++ instrumentation/jedis/jedis-2.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/jedis/v2_0/ShardedJedisClientTest.java
@@ -106,7 +106,9 @@ class ShardedJedisClientTest {
                             equalTo(
                                 SERVER_ADDRESS,
                                 emitStableDatabaseSemconv() ? configuredTarget : shardHost),
-                            equalTo(SERVER_PORT, emitStableDatabaseSemconv() ? null : shardPort))),
+                            equalTo(
+                                SERVER_PORT,
+                                emitStableDatabaseSemconv() ? null : (long) shardPort))),
         trace ->
             trace.hasSpansSatisfyingExactly(
                 span ->
@@ -123,7 +125,9 @@ class ShardedJedisClientTest {
                             equalTo(
                                 SERVER_ADDRESS,
                                 emitStableDatabaseSemconv() ? configuredTarget : shardHost),
-                            equalTo(SERVER_PORT, emitStableDatabaseSemconv() ? null : shardPort))));
+                            equalTo(
+                                SERVER_PORT,
+                                emitStableDatabaseSemconv() ? null : (long) shardPort))));
   }
 
   @Test
@@ -152,6 +156,7 @@ class ShardedJedisClientTest {
                                 SERVER_ADDRESS,
                                 emitStableDatabaseSemconv() ? configuredTarget : selectedHost),
                             equalTo(
-                                SERVER_PORT, emitStableDatabaseSemconv() ? null : selectedPort))));
+                                SERVER_PORT,
+                                emitStableDatabaseSemconv() ? null : (long) selectedPort))));
   }
 }
